- Using <boost/cstdint> via Types.h  - Overall integer type clean up
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1298 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/ChannelServer.cpp
@@ -66,7 +66,7 @@ void ChannelServer::connectWorld() {
 void ChannelServer::loadConfig() {
 	ConfigFile config("conf/channelserver.lua");
 	login_ip = config.getString("login_ip");
-	login_inter_port = config.getInt("login_inter_port");
+	login_inter_port = config.getShort("login_inter_port");
 	external_ip = config.getString("external_ip"); // External IP
 
 	world = -1; // Will get from login server
