+++ src/ChannelServer/FamePacket.h
@@ -18,6 +18,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #ifndef FAMEPACK_H
 #define FAMEPACK_H
 
+#include "Types.h"
 #include <vector>
 
 using std::vector;
@@ -26,7 +27,7 @@ class Player;
 class PacketCreator;
 
 namespace FamePacket{
-	void sendFame(Player *player, Player *player2, int type, int newFame);
-	void sendError(Player *player, int reason);
+	void sendFame(Player *player, Player *player2, uint8_t type, int32_t newFame);
+	void sendError(Player *player, int32_t reason);
 };
 #endif
